[B] refactor: rename package
