+++ src/test/java/lotto/lotto/LottoTicketsTest.java
@@ -1,4 +1,4 @@
-package lotto.ticker;
+package lotto.lotto;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
